Revert "Fix incorrect CPDFSDK_PageView::CountAnnots()."
This reverts commit 9dd9439d23d44d105db0950e33af0cc913a6cece.

BUG=536967
TBR=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1421583005 .

+++ fpdfsdk/include/fsdk_mgr.h
@@ -308,7 +308,7 @@ class CPDFSDK_PageView final {
   CPDFSDK_Annot* AddAnnot(const FX_CHAR* lpSubType, CPDF_Dictionary* pDict);
   CPDFSDK_Annot* AddAnnot(CPDF_Annot* pPDFAnnot);
   FX_BOOL DeleteAnnot(CPDFSDK_Annot* pAnnot);
-  size_t CountAnnots() const;
+  int CountAnnots() const;
   CPDFSDK_Annot* GetAnnot(size_t nIndex);
   CPDFSDK_Annot* GetAnnotByDict(CPDF_Dictionary* pDict);
   CPDF_Page* GetPDFPage() { return m_page; }
